Move ProcessingMessage control to new project structure
Renamed and relocated ProcessingMessage.Designer.cs, ProcessingMessage.cs, and ProcessingMessage.resx from source/main/Controls to src/BowlingMegabucks.TournamentManager.App/Controls. Updated namespaces to reflect the new project structure.
