--- source/main/Controls/ProcessingMessage.cs
@@ -1,31 +1,0 @@
-﻿using System.Runtime.Versioning;
-
-namespace BowlingMegabucks.TournamentManager.Controls;
-
-[SupportedOSPlatform("windows")]
-internal sealed partial class ProcessingMessage : UserControl
-{
-    private readonly CancellationTokenSource _cancelToken;
-
-    /// <summary>
-    ///
-    /// </summary>
-    /// <param name="message"></param>
-    /// <param name="cancelToken"></param>
-    public ProcessingMessage(string message, CancellationTokenSource cancelToken)
-    {
-        InitializeComponent();
-        _cancelToken = cancelToken;
-
-        messageLabel.Text = message;
-
-        if (string.IsNullOrWhiteSpace(message))
-        {
-            messageLabel.Visible = false;
-            processingPicture.Location = new Point(processingPicture.Location.X, processingPicture.Location.Y - (messageLabel.Height / 2));
-        }
-    }
-
-    private void CancelButton_Click(object sender, EventArgs e)
-        => _cancelToken.Cancel();
-}
