+++ src/BowlingMegabucks.TournamentManager.App/Controls/ProcessingMessage.resx
